+++ app/controllers/users/registrations_controller.rb
@@ -49,7 +49,7 @@ class Users::RegistrationsController < Devise::RegistrationsController
 
   # If you have extra params to permit, append them to the sanitizer.
   def configure_account_update_params
-    devise_parameter_sanitizer.permit(:account_update, keys: [:privacy_confirmation])
+    devise_parameter_sanitizer.permit(:account_update, keys: [:accepted_privacy_policy])
   end
 
   def update_resource(resource, params)
