+++ app/helpers/application_helper.rb
@@ -1,19 +1,4 @@
 # frozen_string_literal: true
 
 module ApplicationHelper
-  def resource_name
-    :user
-  end
-
-  def resource
-    @resource ||= User.new
-  end
-
-  def devise_mapping
-    @devise_mapping ||= Devise.mappings[:user]
-  end
-
-  def resource_class
-    User
-  end
 end
