+++ app/helpers/layouts_helper.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module LayoutsHelper
+  def resource_name
+    :user
+  end
+
+  def resource
+    @resource ||= User.new
+  end
+
+  def devise_mapping
+    @devise_mapping ||= Devise.mappings[:user]
+  end
+
+  def resource_class
+    User
+  end
+end
