+++ app/views/layouts/_header.html.erb
@@ -69,7 +69,7 @@
 <%= render partial: "layouts/privacy_confirmation_modal" %>
 
 <% if user_signed_in? %>
-  <% unless current_user.privacy_confirmation %>
+  <% unless current_user.accepted_privacy_policy %>
     <script type="text/javascript">
       $(document).ready(function(){
         if (window.location.pathname != "/privacy" && window.location.pathname != "/tos"){
