+++ db/migrate/20200830082506_change_gdpr_confirmation_column_name.rb
@@ -1,0 +1,5 @@
+class ChangeGdprConfirmationColumnName < ActiveRecord::Migration[6.0]
+  def change
+    rename_column :users, :privacy_confirmation, :accepted_privacy_policy
+  end
+end
