+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_08_21_091915) do
+ActiveRecord::Schema.define(version: 2020_08_30_082506) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -344,7 +344,7 @@ ActiveRecord::Schema.define(version: 2020_08_21_091915) do
     t.string "country"
     t.string "educational_institute"
     t.boolean "subscribed", default: true
-    t.boolean "privacy_confirmation", default: false
+    t.boolean "accepted_privacy_policy", default: false
     t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
